Create aiBasedOffOfFNAF
A model version of FNAF AI where the player spawns at a random location i.e a random number generated then the enemies as well spawn at their random locations. Every interval, a random number is generated and if its greater than or equal to their AI value, that number is added and updates their location. It goes until one of the enemies catches the user